+++ src/module.test.ts
@@ -19,95 +19,34 @@ import { AggregatorEndpoint } from 'matterbridge/matter/endpoints';
 import { WindowCovering, WindowCoveringCluster } from 'matterbridge/matter/clusters';
 
 import initializePlugin, { SomfyTahomaPlatform, SomfyTahomaPlatformConfig } from './module.js';
+import {
+  addBridgedEndpointSpy,
+  addMatterbridgePlatform,
+  createMatterbridgeEnvironment,
+  destroyMatterbridgeEnvironment,
+  log,
+  loggerLogSpy,
+  matterbridge,
+  server,
+  aggregator,
+  setupTest,
+  startMatterbridgeEnvironment,
+  stopMatterbridgeEnvironment,
+  removeAllBridgedEndpointsSpy,
+  setDebug,
+} from './utils/jestHelpers.js';
 
-let loggerLogSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.log>;
-let consoleLogSpy: jest.SpiedFunction<typeof console.log>;
-let consoleDebugSpy: jest.SpiedFunction<typeof console.log>;
-let consoleInfoSpy: jest.SpiedFunction<typeof console.log>;
-let consoleWarnSpy: jest.SpiedFunction<typeof console.log>;
-let consoleErrorSpy: jest.SpiedFunction<typeof console.log>;
-const debug = false; // Set to true to enable debug logs
-
-if (!debug) {
-  loggerLogSpy = jest.spyOn(AnsiLogger.prototype, 'log').mockImplementation((level: string, message: string, ...parameters: any[]) => {});
-  consoleLogSpy = jest.spyOn(console, 'log').mockImplementation((...args: any[]) => {});
-  consoleDebugSpy = jest.spyOn(console, 'debug').mockImplementation((...args: any[]) => {});
-  consoleInfoSpy = jest.spyOn(console, 'info').mockImplementation((...args: any[]) => {});
-  consoleWarnSpy = jest.spyOn(console, 'warn').mockImplementation((...args: any[]) => {});
-  consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation((...args: any[]) => {});
-} else {
-  loggerLogSpy = jest.spyOn(AnsiLogger.prototype, 'log');
-  consoleLogSpy = jest.spyOn(console, 'log');
-  consoleDebugSpy = jest.spyOn(console, 'debug');
-  consoleInfoSpy = jest.spyOn(console, 'info');
-  consoleWarnSpy = jest.spyOn(console, 'warn');
-  consoleErrorSpy = jest.spyOn(console, 'error');
-}
-
-// Cleanup the test environment
-rmSync(HOMEDIR, { recursive: true, force: true });
+// Setup the test environment
+setupTest(NAME, false);
 
 describe('TestPlatform', () => {
-  let matterbridge: Matterbridge;
-  let server: ServerNode<ServerNode.RootEndpoint>;
-  let aggregator: Endpoint<AggregatorEndpoint>;
-  let device: MatterbridgeEndpoint;
-
   let somfyPlatform: SomfyTahomaPlatform;
 
   let clientConnectSpy: jest.SpiedFunction<(user: string, password: string) => Promise<void>>;
   let clientGetDevicesSpy: jest.SpiedFunction<() => Promise<Device[]>>;
   let clientExecuteSpy: jest.SpiedFunction<(oid: any, execution: any) => Promise<any>>;
 
-  const mockLog = {
-    fatal: jest.fn((message: string, ...parameters: any[]) => {
-      // console.log('mockLog.fatal', message, parameters);
-    }),
-    error: jest.fn((message: string, ...parameters: any[]) => {
-      // console.log('mockLog.error', message, parameters);
-    }),
-    warn: jest.fn((message: string, ...parameters: any[]) => {
-      // console.log('mockLog.warn', message, parameters);
-    }),
-    notice: jest.fn((message: string, ...parameters: any[]) => {
-      // console.log('mockLog.notice', message, parameters);
-    }),
-    info: jest.fn((message: string, ...parameters: any[]) => {
-      // console.log('mockLog.info', message, parameters);
-    }),
-    debug: jest.fn((message: string, ...parameters: any[]) => {
-      // console.log('mockLog.debug', message, parameters);
-    }),
-  } as unknown as AnsiLogger;
-
-  const mockMatterbridge = {
-    homeDirectory: path.join(HOMEDIR),
-    matterbridgeDirectory: path.join(HOMEDIR, '.matterbridge'),
-    matterbridgePluginDirectory: path.join(HOMEDIR, 'Matterbridge'),
-    systemInformation: { ipv4Address: undefined, ipv6Address: undefined, osRelease: 'xx.xx.xx.xx.xx.xx', nodeVersion: '22.1.10' },
-    matterbridgeVersion: '3.3.0',
-    log: mockLog,
-    getDevices: jest.fn(() => {
-      // console.log('getDevices called');
-      return [];
-    }),
-    getPlugins: jest.fn(() => {
-      // console.log('getDevices called');
-      return [];
-    }),
-    addBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
-      console.log('addBridgedEndpoint called');
-      await aggregator.add(device);
-    }),
-    removeBridgedEndpoint: jest.fn(async (pluginName: string, device: MatterbridgeEndpoint) => {
-      // console.log('removeBridgedEndpoint called');
-    }),
-    removeAllBridgedEndpoints: jest.fn(async (pluginName: string) => {
-      // console.log('removeAllBridgedEndpoints called');
-    }),
-  } as unknown as Matterbridge;
-
-  const mockConfig: SomfyTahomaPlatformConfig = {
+  const config: SomfyTahomaPlatformConfig = {
     name: 'matterbridge-somfy-tahoma',
     type: 'DynamicPlatform',
     version: '1.4.0',
@@ -136,19 +75,9 @@ describe('TestPlatform', () => {
   ];
 
   beforeAll(async () => {
-    // Create a MatterbridgeEdge instance
-    matterbridge = await Matterbridge.loadInstance(false);
-    matterbridge.homeDirectory = path.join(HOMEDIR);
-    matterbridge.matterbridgeDirectory = path.join(HOMEDIR, '.matterbridge');
-    matterbridge.matterbridgePluginDirectory = path.join(HOMEDIR, 'Matterbridge');
-
-    // Setup matter environment
-    matterbridge.environment.vars.set('log.level', Level.INFO);
-    matterbridge.environment.vars.set('log.format', Format.ANSI);
-    matterbridge.environment.vars.set('path.root', path.join(HOMEDIR, '.matterbridge', 'matterstorage'));
-    matterbridge.environment.vars.set('runtime.signals', false);
-    matterbridge.environment.vars.set('runtime.exitcode', false);
-    await (matterbridge as any).startMatterStorage();
+    // Create Matterbridge environment
+    await createMatterbridgeEnvironment(NAME);
+    await startMatterbridgeEnvironment(MATTER_PORT);
 
     // Spy on the Client.connect method
     clientConnectSpy = jest.spyOn(Client.prototype, 'connect').mockImplementation((user: string, password: string) => {
@@ -166,85 +95,63 @@ describe('TestPlatform', () => {
   });
 
   beforeEach(async () => {
-    // Close the Matterbridge instance
-    await matterbridge.destroyInstance();
-
-    // Restore all mocks
+    // Reset the mock calls before each test
     jest.clearAllMocks();
   });
 
-  it('should create the storage context', async () => {
-    await (matterbridge as any).startMatterStorage();
-    expect(matterbridge.matterStorageService).toBeDefined();
-    expect(matterbridge.matterStorageManager).toBeDefined();
-    expect(matterbridge.matterbridgeContext).toBeDefined();
-  });
+  afterEach(async () => {});
 
-  it('should create the server', async () => {
-    server = await (matterbridge as any).createServerNode(matterbridge.matterbridgeContext);
-    expect(server).toBeDefined();
-  });
+  afterAll(async () => {
+    // Destroy Matterbridge environment
+    await stopMatterbridgeEnvironment();
+    await destroyMatterbridgeEnvironment();
 
-  it('should create the aggregator', async () => {
-    aggregator = await (matterbridge as any).createAggregatorNode(matterbridge.matterbridgeContext);
-    expect(aggregator).toBeDefined();
-  });
-
-  it('should add the aggregator to the server', async () => {
-    expect(await server.add(aggregator)).toBeDefined();
-  });
-
-  it('should start the server', async () => {
-    await (matterbridge as any).startServerNode(server);
-    expect(server).toBeDefined();
-    expect(server?.lifecycle.isReady).toBeTruthy();
-    expect(server?.construction.status).toBe(Lifecycle.Status.Active);
-    expect(aggregator).toBeDefined();
-    expect(aggregator?.lifecycle.isReady).toBeTruthy();
-    expect(aggregator?.construction.status).toBe(Lifecycle.Status.Active);
+    // Restore all mocks
+    jest.restoreAllMocks();
   });
 
   it('should return an instance of SomfyTahomaPlatform', async () => {
-    const result = initializePlugin(mockMatterbridge, mockLog, mockConfig);
+    const result = initializePlugin(matterbridge, log, config);
     expect(result).toBeInstanceOf(SomfyTahomaPlatform);
     await result.onShutdown();
   });
 
   it('should not initialize platform without username and password', () => {
-    mockConfig.username = '';
-    mockConfig.password = '';
-    mockConfig.service = '';
-    somfyPlatform = new SomfyTahomaPlatform(mockMatterbridge, mockLog, mockConfig);
-    expect(mockLog.info).toHaveBeenCalledWith('Initializing platform:', mockConfig.name);
-    expect(mockLog.error).toHaveBeenCalledWith('No service or username or password provided for:', mockConfig.name);
+    config.username = '';
+    config.password = '';
+    config.service = '';
+    somfyPlatform = new SomfyTahomaPlatform(matterbridge, log, config);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'Initializing platform:', config.name);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, 'No service or username or password provided for:', config.name);
   });
 
   it('should initialize platform with config name', () => {
-    mockConfig.username = 'None';
-    mockConfig.password = 'None';
-    mockConfig.service = 'somfy_europe';
-    somfyPlatform = new SomfyTahomaPlatform(mockMatterbridge, mockLog, mockConfig);
-    expect(mockLog.info).toHaveBeenCalledWith('Initializing platform:', mockConfig.name);
-    expect(mockLog.info).toHaveBeenCalledWith('Finished initializing platform:', mockConfig.name);
-    expect(mockLog.info).toHaveBeenCalledWith('Starting client Tahoma service somfy_europe with user None password: None');
+    config.username = 'None';
+    config.password = 'None';
+    config.service = 'somfy_europe';
+    somfyPlatform = new SomfyTahomaPlatform(matterbridge, log, config);
+    addMatterbridgePlatform(somfyPlatform, 'matterbridge-somfy-tahoma');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'Initializing platform:', config.name);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'Finished initializing platform:', config.name);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'Starting client Tahoma service somfy_europe with user None password: None');
   });
 
   it('should receive tahomaClient events', () => {
     (somfyPlatform as any).tahomaClient?.emit('connect');
     (somfyPlatform as any).tahomaClient?.emit('disconnect');
-    expect(mockLog.info).toHaveBeenCalledWith('TaHoma service connected');
-    expect(mockLog.warn).toHaveBeenCalledWith('TaHoma service disconnected');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'TaHoma service connected');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.WARN, 'TaHoma service disconnected');
   });
 
   it('should throw because of version', () => {
-    mockMatterbridge.matterbridgeVersion = '1.5.4';
-    expect(() => new SomfyTahomaPlatform(mockMatterbridge, mockLog, mockConfig)).toThrow();
-    mockMatterbridge.matterbridgeVersion = '3.3.0';
+    matterbridge.matterbridgeVersion = '1.5.4';
+    expect(() => new SomfyTahomaPlatform(matterbridge, log, config)).toThrow();
+    matterbridge.matterbridgeVersion = '3.3.0';
   });
 
   it('should call onStart with reason', async () => {
     await somfyPlatform.onStart('Test reason');
-    expect(mockLog.info).toHaveBeenCalledWith('onStart called with reason:', 'Test reason');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onStart called with reason:', 'Test reason');
     expect(clientConnectSpy).toHaveBeenCalledWith('None', 'None');
   });
 
@@ -252,7 +159,7 @@ describe('TestPlatform', () => {
     const client = (somfyPlatform as any).tahomaClient;
     (somfyPlatform as any).tahomaClient = undefined;
     await somfyPlatform.onStart();
-    expect(mockLog.error).toHaveBeenCalledWith('TaHoma service not created');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, 'TaHoma service not created');
     expect(clientConnectSpy).not.toHaveBeenCalledWith('None', 'None');
     (somfyPlatform as any).tahomaClient = client;
   });
@@ -262,9 +169,9 @@ describe('TestPlatform', () => {
       throw new Error('Error connecting to TaHoma service');
     });
     await somfyPlatform.onStart('Test reason');
-    expect(mockLog.info).toHaveBeenCalledWith('onStart called with reason:', 'Test reason');
-    expect(mockLog.error).not.toHaveBeenCalledWith('TaHoma service not created');
-    expect(mockLog.error).toHaveBeenCalledWith(expect.stringContaining('Error connecting to TaHoma service'), undefined);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onStart called with reason:', 'Test reason');
+    expect(loggerLogSpy).not.toHaveBeenCalledWith(LogLevel.ERROR, 'TaHoma service not created');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.stringContaining('Error connecting to TaHoma service'));
     expect(clientConnectSpy).toHaveBeenCalledWith('None', 'None');
   });
 
@@ -272,17 +179,17 @@ describe('TestPlatform', () => {
     const client = (somfyPlatform as any).tahomaClient;
     (somfyPlatform as any).tahomaClient = undefined;
     await (somfyPlatform as any).discoverDevices();
-    expect(mockLog.error).toHaveBeenCalledWith('TaHoma service not created');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, 'TaHoma service not created');
     (somfyPlatform as any).tahomaClient = client;
   });
 
   it('should log an error if writeFile fails', async () => {
-    const fileName = path.join(mockMatterbridge.matterbridgePluginDirectory, 'matterbridge-somfy-tahoma', 'devices.json');
+    const fileName = path.join(matterbridge.matterbridgePluginDirectory, 'matterbridge-somfy-tahoma', 'devices.json');
     const errorMessage = 'Error writing file';
     jest.spyOn(fs, 'writeFile').mockRejectedValueOnce(new Error(errorMessage));
     await (somfyPlatform as any).discoverDevices();
     await wait(1000);
-    expect(mockLog.error).toHaveBeenCalled();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.anything());
   });
 
   it('should discover devices and log error if getDevices throws', async () => {
@@ -290,7 +197,7 @@ describe('TestPlatform', () => {
       throw new Error('Error getting devices from TaHoma service');
     });
     await (somfyPlatform as any).discoverDevices();
-    expect(mockLog.error).toHaveBeenCalledWith('Error discovering TaHoma devices:', undefined);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.stringContaining('Error discovering TaHoma devices'));
   });
 
   it('should discover devices and not add if in black list', async () => {
@@ -299,9 +206,10 @@ describe('TestPlatform', () => {
       return Promise.resolve(mockDevices);
     });
     await (somfyPlatform as any).discoverDevices();
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma devices`);
+    expect(addBridgedEndpointSpy).toHaveBeenCalledTimes(0);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma devices`);
     expect((somfyPlatform as any).tahomaDevices).toHaveLength(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma screens`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma screens`);
     expect((somfyPlatform as any).bridgedDevices).toHaveLength(0);
     expect((somfyPlatform as any).covers.size).toBe(0);
     somfyPlatform.config.blackList = [];
@@ -316,12 +224,12 @@ describe('TestPlatform', () => {
       return Promise.resolve(mockDevices);
     });
     await (somfyPlatform as any).discoverDevices();
-    expect(mockMatterbridge.addBridgedEndpoint).toHaveBeenCalledTimes(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma devices`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`- added with uniqueName`);
+    expect(addBridgedEndpointSpy).toHaveBeenCalledTimes(1);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma devices`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `- added with uniqueName`);
     expect((somfyPlatform as any).tahomaDevices).toHaveLength(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma screens`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma screens`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
     expect((somfyPlatform as any).bridgedDevices).toHaveLength(1);
     expect((somfyPlatform as any).covers.size).toBe(1);
     console.log('Deleting device');
@@ -330,6 +238,9 @@ describe('TestPlatform', () => {
     (somfyPlatform as any).bridgedDevices = [];
     (somfyPlatform as any).covers.clear();
     (somfyPlatform as any).registeredEndpointsByName.clear();
+    matterbridge.devices.clear();
+    expect(aggregator.parts.size).toBe(0);
+    expect(matterbridge.devices.size).toBe(0);
   });
 
   it('should discover devices with uiClass Screen', async () => {
@@ -339,12 +250,12 @@ describe('TestPlatform', () => {
       return Promise.resolve(mockDevices);
     });
     await (somfyPlatform as any).discoverDevices();
-    expect(mockMatterbridge.addBridgedEndpoint).toHaveBeenCalledTimes(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma devices`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`- added with uiClass`);
+    expect(addBridgedEndpointSpy).toHaveBeenCalledTimes(1);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma devices`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `- added with uiClass`);
     expect((somfyPlatform as any).tahomaDevices).toHaveLength(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma screens`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma screens`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
     expect((somfyPlatform as any).bridgedDevices).toHaveLength(1);
     expect((somfyPlatform as any).covers.size).toBe(1);
     console.log('Deleting device');
@@ -353,6 +264,9 @@ describe('TestPlatform', () => {
     (somfyPlatform as any).bridgedDevices = [];
     (somfyPlatform as any).covers.clear();
     (somfyPlatform as any).registeredEndpointsByName.clear();
+    matterbridge.devices.clear();
+    expect(aggregator.parts.size).toBe(0);
+    expect(matterbridge.devices.size).toBe(0);
   });
 
   it('should discover devices with command "open", "close" and "stop"', async () => {
@@ -362,28 +276,28 @@ describe('TestPlatform', () => {
       return Promise.resolve(mockDevices);
     });
     await (somfyPlatform as any).discoverDevices();
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma devices`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`- added with commands "open", "close" and "stop"`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma devices`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `- added with commands "open", "close" and "stop"`);
     expect((somfyPlatform as any).tahomaDevices).toHaveLength(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma screens`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma screens`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
     expect((somfyPlatform as any).bridgedDevices).toHaveLength(1);
     expect((somfyPlatform as any).covers.size).toBe(1);
 
     (somfyPlatform as any).sendCommand('identify', mockDevices[0]);
-    expect(mockLog.info).toHaveBeenCalledWith(`Sending command ${YELLOW}identify${nf} highPriority false`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Sending command ${YELLOW}identify${nf} highPriority false`);
     (somfyPlatform as any).sendCommand('open', mockDevices[0]);
-    expect(mockLog.info).toHaveBeenCalledWith(`Sending command ${YELLOW}open${nf} highPriority false`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Sending command ${YELLOW}open${nf} highPriority false`);
     (somfyPlatform as any).sendCommand('stop', mockDevices[0]);
-    expect(mockLog.info).toHaveBeenCalledWith(`Sending command ${YELLOW}stop${nf} highPriority false`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Sending command ${YELLOW}stop${nf} highPriority false`);
     (somfyPlatform as any).sendCommand('close', mockDevices[0]);
-    expect(mockLog.info).toHaveBeenCalledWith(`Sending command ${YELLOW}close${nf} highPriority false`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Sending command ${YELLOW}close${nf} highPriority false`);
 
     clientExecuteSpy.mockImplementationOnce(() => {
       throw new Error('Error executing command');
     });
     (somfyPlatform as any).sendCommand('close', mockDevices[0]);
-    expect(mockLog.error).toHaveBeenCalledWith(expect.stringContaining(`Error sending command`));
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, expect.stringContaining(`Error sending command`));
 
     const device = somfyPlatform.covers.get('Device1')?.bridgedDevice;
     expect(device).toBeDefined();
@@ -449,6 +363,9 @@ describe('TestPlatform', () => {
     (somfyPlatform as any).bridgedDevices = [];
     (somfyPlatform as any).covers.clear();
     (somfyPlatform as any).registeredEndpointsByName.clear();
+    matterbridge.devices.clear();
+    expect(aggregator.parts.size).toBe(0);
+    expect(matterbridge.devices.size).toBe(0);
   }, 120000);
 
   it('should discover devices with command "rollOut", "rollUp" and "stop"', async () => {
@@ -459,11 +376,11 @@ describe('TestPlatform', () => {
       return Promise.resolve(mockDevices);
     });
     await (somfyPlatform as any).discoverDevices();
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma devices`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`- added with commands "rollOut", "rollUp" and "stop"`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma devices`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `- added with commands "rollOut", "rollUp" and "stop"`);
     expect((somfyPlatform as any).tahomaDevices).toHaveLength(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma screens`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma screens`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
     expect((somfyPlatform as any).bridgedDevices).toHaveLength(1);
     expect((somfyPlatform as any).covers.size).toBe(1);
     (somfyPlatform as any).sendCommand('identify', mockDevices[0]);
@@ -475,6 +392,9 @@ describe('TestPlatform', () => {
     (somfyPlatform as any).bridgedDevices = [];
     (somfyPlatform as any).covers.clear();
     (somfyPlatform as any).registeredEndpointsByName.clear();
+    matterbridge.devices.clear();
+    expect(aggregator.parts.size).toBe(0);
+    expect(matterbridge.devices.size).toBe(0);
   });
 
   it('should discover devices with command "down", "up" and "stop"', async () => {
@@ -485,70 +405,66 @@ describe('TestPlatform', () => {
       return Promise.resolve(mockDevices);
     });
     await (somfyPlatform as any).discoverDevices();
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma devices`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`- added with commands "down", "up" and "stop"`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma devices`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `- added with commands "down", "up" and "stop"`);
     expect((somfyPlatform as any).tahomaDevices).toHaveLength(1);
-    expect(mockLog.info).toHaveBeenCalledWith(`Discovered 1 TaHoma screens`);
-    expect(mockLog.debug).toHaveBeenCalledWith(`Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, `Discovered 1 TaHoma screens`);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.DEBUG, `Adding device: ${BLUE}${mockDevices[0].label}${rs}`);
     expect((somfyPlatform as any).bridgedDevices).toHaveLength(1);
     expect((somfyPlatform as any).covers.size).toBe(1);
     (somfyPlatform as any).sendCommand('identify', mockDevices[0]);
     (somfyPlatform as any).sendCommand('open', mockDevices[0]);
     (somfyPlatform as any).sendCommand('stop', mockDevices[0]);
     (somfyPlatform as any).sendCommand('close', mockDevices[0]);
+    expect(aggregator.parts.size).toBe(1);
+    expect(matterbridge.devices.size).toBe(1);
+    // We keep this device to be used in the next tests
   });
 
   it('should call onConfigure', async () => {
     await somfyPlatform.onConfigure();
-    expect(mockLog.info).toHaveBeenCalledWith('onConfigure called');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onConfigure called');
   });
 
   it('should call onConfigure and log error', async () => {
     const client = (somfyPlatform as any).tahomaClient;
     (somfyPlatform as any).tahomaClient = undefined;
     await somfyPlatform.onConfigure();
-    expect(mockLog.info).toHaveBeenCalledWith('onConfigure called');
-    expect(mockLog.error).toHaveBeenCalledWith('TaHoma service not created');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onConfigure called');
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.ERROR, 'TaHoma service not created');
     (somfyPlatform as any).tahomaClient = client;
   });
 
   it('should call onShutdown with reason', async () => {
+    expect(aggregator.parts.size).toBe(1);
+    setDebug(true);
     const client = (somfyPlatform as any).tahomaClient;
     (somfyPlatform as any).tahomaClient = undefined;
     await somfyPlatform.onShutdown('Test reason');
-    expect(mockLog.info).toHaveBeenCalledWith('onShutdown called with reason:', 'Test reason');
-    expect((somfyPlatform as any).tahomaClient).toBeUndefined();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onShutdown called with reason:', 'Test reason');
     (somfyPlatform as any).tahomaClient = client;
+    expect(aggregator.parts.size).toBe(1);
+    expect(matterbridge.devices.size).toBe(1);
+    expect(removeAllBridgedEndpointsSpy).toHaveBeenCalledTimes(0);
   });
 
   it('should call onShutdown with reason and call unregisterAll', async () => {
     const client = (somfyPlatform as any).tahomaClient;
     (somfyPlatform as any).tahomaClient = undefined;
-    somfyPlatform.name = mockConfig.name as string;
-    mockConfig.unregisterOnShutdown = true;
+    somfyPlatform.name = config.name as string;
+    config.unregisterOnShutdown = true;
     await somfyPlatform.onShutdown();
-    expect(mockLog.info).toHaveBeenCalledWith('onShutdown called with reason:', 'none');
-    expect(mockMatterbridge.removeAllBridgedEndpoints).toHaveBeenCalledWith(mockConfig.name, 0);
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onShutdown called with reason:', 'none');
+    expect(removeAllBridgedEndpointsSpy).toHaveBeenCalledWith(config.name, 0);
     expect((somfyPlatform as any).tahomaClient).toBeUndefined();
     (somfyPlatform as any).tahomaClient = client;
-    mockConfig.unregisterOnShutdown = false;
+    config.unregisterOnShutdown = false;
+    expect(aggregator.parts.size).toBe(0);
+    expect(matterbridge.devices.size).toBe(0);
   });
 
   it('should call onShutdown with reason and log error', async () => {
     await somfyPlatform.onShutdown('Test reason');
-    expect(mockLog.info).toHaveBeenCalledWith('onShutdown called with reason:', 'Test reason');
-  });
-
-  it('should stop the server', async () => {
-    await (matterbridge as any).stopServerNode(server);
-    expect(server.lifecycle.isOnline).toBe(false);
-    await server.env.get(MdnsService)[Symbol.asyncDispose]();
-  });
-
-  it('should stop the storage', async () => {
-    await (matterbridge as any).stopMatterStorage();
-    expect(matterbridge.matterStorageService).not.toBeDefined();
-    expect(matterbridge.matterStorageManager).not.toBeDefined();
-    expect(matterbridge.matterbridgeContext).not.toBeDefined();
+    expect(loggerLogSpy).toHaveBeenCalledWith(LogLevel.INFO, 'onShutdown called with reason:', 'Test reason');
   });
 });